app/pyproject.toml
[project]
name = "mongo-db-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "mcp>=1.9.0",
    "pydantic-settings>=2.8.1",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
    "fastmcp>=2.8.1",
    "bcrypt>=4.3.0",
    "passlib>=1.7.4",
    "python-jose>=3.4.0",
    "fastapi-mcp>=0.4.0",
]
